requirements.txt
asgiref==3.8.1
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.6
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==5.2.2
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pytesseract==0.3.10
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.0
urllib3==2.2.2
